+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.15.0] - 2023-07-28
+### Added
+- [@mjkuranda](https://github.com/mjkuranda): MenuOptionValue, StringMenuOptionValue to store form values.
+- [@mjkuranda](https://github.com/mjkuranda): InputMenuOptionEvent to handling inputting keys.
+- [@mjkuranda](https://github.com/mjkuranda): InputMenuOption as formant to inputting string.
+
 ### Changed
 - [@mjkuranda](https://github.com/mjkuranda): GameOverState menu state.